requirements.txt
instagrapi==1.12.0
Pillow==10.3.0
Flask==3.0.3
python-dotenv
requests
